requirements.txt
aniso8601==7.0.0
asgiref==3.3.1
attrs==20.3.0
Django==3.1.6
django-filter==2.4.0
graphene==2.1.8
graphene-django==2.15.0
graphene-federation==0.1.0
graphql-core==2.3.2
graphql-relay==2.0.1
iniconfig==1.1.1
opentracing==2.4.0
packaging==20.9
pkg-resources==0.0.0
pluggy==0.13.1
promise==2.3
py==1.10.0
pyparsing==2.4.7
pytest==6.2.2
pytz==2021.1
Rx==1.6.1
singledispatch==3.4.0.3
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
